+++ public/css/watch.css
@@ -29,28 +29,28 @@ a:hover {
   text-decoration: underline;
 }
 
-.divider {  
-    border-top: 3px solid var(--label-bg-color);  
+.divider {
+  border-top: 3px solid var(--label-bg-color);
 }
 
 /* Icon */
 .icon--arrow-down {
-  background: url("../images/icon-arrow-down.png");
+  background: url('../images/icon-arrow-down.png');
 }
 .icon--arrow-up {
-  background: url("../images/icon-arrow-up.png");
+  background: url('../images/icon-arrow-up.png');
 }
 .icon--notice {
   width: 9px;
   height: 9px;
   margin: 0 5px;
-  background: url("../images/icon-red-star.gif");  
+  background: url('../images/icon-red-star.gif');
 }
 
 /* Title */
 .title {
   height: 178px;
-  background: url("../images/bg-ms.jpg");
+  background: url('../images/bg-ms.jpg');
 }
 .title .title__container {
   width: 70%;
@@ -89,7 +89,7 @@ a:hover {
 }
 .content__suggestion {
   border-radius: 3px;
-  background: #f4f4f2 url("../images/bg-suggestion2.png") 540px bottom no-repeat;
+  background: #f4f4f2 url('../images/bg-suggestion2.png') 540px bottom no-repeat;
   padding: 30px;
   margin-bottom: 30px;
 }
@@ -99,7 +99,7 @@ a:hover {
   color: var(--black-color);
 }
 /* li 2번부터 4번까지 */
-.content__suggestion ul li:nth-child(n+2):nth-child(-n+4) {
+.content__suggestion ul li:nth-child(n + 2):nth-child(-n + 4) {
   font-size: var(--font-medium);
   color: var(--black-grey-color);
 }
@@ -108,11 +108,11 @@ a:hover {
   font-size: var(--font-small);
   color: var(--grey-color);
 }
-.content__suggetion__btn {
+.content__suggestion__btn {
   display: flex;
   justify-content: flex-start;
 }
-.suggetion-btn {
+.suggestion-btn {
   font-weight: bold;
   font-size: var(--font-small);
   color: var(--black-color);
@@ -122,14 +122,14 @@ a:hover {
   margin: 10px 10px 10px 0;
   padding: 10px 20px;
 }
-.suggetion-btn:hover {
+.suggestion-btn:hover {
   cursor: pointer;
   text-decoration: underline;
 }
 
 /* Form */
 .`__notice {
-  display: flex;  
+  display: flex;
   align-items: center;
   justify-content: flex-end;
 }
@@ -140,16 +140,16 @@ a:hover {
 }
 .content__form__row {
   display: flex;
-  height: 47px;  
+  height: 47px;
   border-bottom: 1px solid #e0e0e0;
 }
-.content__form__row.first{
+.content__form__row.first {
   border-top: 1px solid var(--black-color);
 }
-.content__form__row.form--content{
-  height: 221px;  
+.content__form__row.form--content {
+  height: 221px;
 }
-.content__form__row.form--content.answer{
+.content__form__row.form--content.answer {
   height: 221px;
   border-top: 1px solid #e0e0e0;
 }
@@ -171,7 +171,7 @@ a:hover {
 }
 .content__form__row input {
   border: 1px solid var(--input-border-color);
-  border-radius: 3px;  
+  border-radius: 3px;
   height: 26px;
   padding: 0;
 }
@@ -197,11 +197,11 @@ a:hover {
 }
 .phone-select {
   width: var(--input-small);
-  height: 28px;  
+  height: 28px;
 }
 .hyphen {
-  color: var(--black-grey-color);    
-  font-size: var(--font-large);    
+  color: var(--black-grey-color);
+  font-size: var(--font-large);
   width: 20px;
   text-align: center;
 }
@@ -210,7 +210,7 @@ a:hover {
 }
 .content__form__row__btn {
   display: flex;
-  justify-content: flex-end;  
+  justify-content: flex-end;
 }
 /* Navigation */
 .content > nav {
