+++ views/edit.pug
@@ -23,8 +23,8 @@ block content
           li 스타벅스에서는 고객 여러분의 즐겁고 행복한 시간을 위해 정성을 다하고 있습니다.
           li 만족스러운 서비스였는지, 불만스러운 점은 없으셨는지 귀한 의견을 들려주시기 바랍니다.          
           li 고객의 소리 운영시간 : 09:00 ~ 18:00 (연중 무휴)        
-        div.content__suggetion__btn
-          a(href="/user/my").suggetion-btn 나의 문의 내역 보기
+        div.content__suggestion__btn
+          a(href="/user/my").suggestion-btn 나의 문의 내역 보기
       section.content__form
         div.content__form__notice
           span.icon--notice
@@ -77,16 +77,16 @@ block content
               label(for="title") 제목
               span.icon--notice
             div.content__form__row__input-container
-              input(type="text" name="title" id="title" required value=suggetion.title).form-title.input-large
+              input(type="text" name="title" id="title" required value=suggestion.title).form-title.input-large
           div.content__form__row.form--content
             div.content__form__row-title
               label(for="content") 내용
               span.icon--notice
             div.content__form__row__input-container
-              input(type="text" name="content" id="content" required value=suggetion.content).form-content.input-large
+              input(type="text" name="content" id="content" required value=suggestion.content).form-content.input-large
           if !user.isAdmin
             div.content__form__row__btn
-              input(type="submit" value="고객의 소리 변경하기").suggetion-btn
+              input(type="submit" value="고객의 소리 변경하기").suggestion-btn
           if user.isAdmin 
             hr.divider
             section.content__answer
@@ -98,7 +98,7 @@ block content
                   //- 답변 내용 받아와서 여기 input의 value 속성값에 넣어주세요😀
                   input(type="text" name="answer" id="answer" value=answer.contents required).form-content.input-large
               div.content__form__row__btn
-                button(data-answer=answer._id, data-suggestion=suggetion._id).suggetion-btn.suggetion-btn__edit 수정하기
+                button(data-answer=answer._id, data-suggestion=suggestion._id).suggestion-btn.suggestion-btn__edit 수정하기
     nav
       ul
         li
